refactor: PCA package
