+++ src/main/java/org/rsultan/example/PCAExample.java
@@ -4,7 +4,7 @@ import java.io.IOException;
 import org.nd4j.linalg.api.buffer.DataType;
 import org.nd4j.linalg.factory.Nd4j;
 import org.nd4j.linalg.ops.transforms.Transforms;
-import org.rsultan.core.classification.PrincipalComponentAnalysis;
+import org.rsultan.core.dimred.PrincipalComponentAnalysis;
 import org.rsultan.dataframe.Dataframes;
 
 public class PCAExample {
